Update patch set 1
Patch Set 1:

I wiped and tested it several times on my end and it seems to work perfectly.

After google credentials are entered, it launches the app allowing the user to choose applications desired.  Once the user clicks "install" on the market intents, the application automatically disables itself for good and setupwizard continues.

Patch-set: 1